tidy up SEvt::meta collection, avoid playing pass the parcel, add C4Version and uname to SEvt metadata

+++ examples/UseCustom4/go.sh
@@ -13,26 +13,40 @@ To remove all installed Custom4 libs and headers::
 EOU
 }
 
-echo $CMAKE_PREFIX_PATH | tr ":" "\n"
 
-sdir=$(pwd)
+sdir=$(cd $(dirname $BASH_SOURCE) && pwd)
 name=$(basename $sdir)
 bdir=/tmp/$USER/opticks/$name/build 
 idir=/tmp/$USER/opticks/$name/install
-
-rm -rf $idir && mkdir -p $idir && cd $idir && pwd 
-rm -rf $bdir && mkdir -p $bdir && cd $bdir && pwd 
-
-cmake $sdir -DCMAKE_BUILD_TYPE=Debug -DCMAKE_INSTALL_PREFIX=$idir 
-make
-make install   
-
 bin=$idir/lib/$name
+vars="sdir name bdir idir bin"
 
-echo $BASH_SOURCE : bin $bin : running 
+defarg="info_build_run"
+arg=${1:-$defarg}
 
-$bin
-rc=$?
+if [ "${arg/info}" != "$arg" ]; then 
+   echo $CMAKE_PREFIX_PATH | tr ":" "\n"
+   for var in $vars ; do printf "%20s : %s\n" "$var" "${!var}" ; done 
+fi 
 
-echo $BASH_SOURCE : bin $bin : rc $rc 
+if [ "${arg/build}" != "$arg" ]; then 
+
+    rm -rf $idir && mkdir -p $idir && cd $idir && pwd 
+    rm -rf $bdir && mkdir -p $bdir && cd $bdir && pwd 
+
+    cmake $sdir -DCMAKE_BUILD_TYPE=Debug -DCMAKE_INSTALL_PREFIX=$idir 
+    [ $? -ne 0 ] && echo $BASH_SOURCE : build/cmake error && exit 1 
+    make
+    [ $? -ne 0 ] && echo $BASH_SOURCE : build/make error && exit 1 
+    make install   
+    [ $? -ne 0 ] && echo $BASH_SOURCE : build/install error && exit 1 
+fi 
+
+
+if [ "${arg/run}" != "$arg" ]; then 
+    $bin
+    [ $? -ne 0 ] && echo $BASH_SOURCE : run error && exit 2
+fi 
+
+exit 0
